refactor utils.py & multiple house sim api
