+++ pysimdeum/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 
-from .api import built_house
+from .api import built_house, build_multi_hh
