+++ pysimdeum/core/user.py
@@ -2,7 +2,7 @@ import copy
 import numpy as np
 import scipy.stats as sstats
 import pandas as pd
-from pysimdeum.core.utils import Base
+from pysimdeum.utils.base import Base
 from dataclasses import dataclass, field
 from typing import Any, Callable, Literal
 from pysimdeum.core.statistics import Statistics
